+++ src/point.rs
@@ -3,13 +3,13 @@ use std::ops::{Add, Sub};
 
 #[derive(Copy, Clone, Debug, Default)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 impl Point {
-    pub fn new(x: f32, y: f32, z: f32) -> Point {
+    pub fn new(x: f64, y: f64, z: f64) -> Point {
         Point { x, y, z }
     }
 }
